feat: export vite preprocessors
