+++ packages/e2e-tests/preprocess-with-vite/svelte.config.js
@@ -1,0 +1,6 @@
+// eslint-disable-next-line node/no-missing-require
+const { viteScript, viteStyle } = require('@sveltejs/vite-plugin-svelte/preprocess');
+
+module.exports = {
+	preprocess: [viteScript(), viteStyle()]
+};
