+++ packages/e2e-tests/preprocess-with-vite/vite.config.js
@@ -4,13 +4,7 @@ const { defineConfig } = require('vite');
 module.exports = defineConfig(({ command, mode }) => {
 	const isProduction = mode === 'production';
 	return {
-		plugins: [
-			svelte({
-				experimental: {
-					useVitePreprocess: true
-				}
-			})
-		],
+		plugins: [svelte()],
 		build: {
 			minify: isProduction
 		}
